refactor: moved all newly private TCP headers into private include folder
